Refactor org views tests into services tests
